refactor(editor): rename MarkdownViewToggle to EditorViewToggle
The toggle now drives both markdown sub-modes and the Edit/Changes
segmented control for plain code files, so the "markdown" in the name
no longer matches its scope. Also shorten the Changes hover tooltip
from "Changes — uncommitted changes (working tree vs HEAD)" to
"Uncommitted changes" — the button label already says "Changes," so the
tooltip's job is to disambiguate, not repeat.
